src/main/java/com/example/endpoint/transaction/resource/TransactionResource.java
package com.example.endpoint.transaction.resource;

import com.example.endpoint.transaction.model.Transaction;
import com.example.endpoint.transaction.service.TransactionService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.MediaType;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

import java.util.List;

@RestController
@RequestMapping("/transactions")
public class TransactionResource {

    @Autowired
    private TransactionService transactionService;

    @GetMapping(consumes = MediaType.APPLICATION_JSON_VALUE, produces = MediaType.APPLICATION_JSON_VALUE)
    public List<Transaction> getTransactions() {
        return transactionService.getTransactions();
    }
}
